fix: adjust file id shown in details page
In the details page the pdf id was shown as the file id. This was not correct
as the file name was made up of the user id and a generated uuid. Now the
correct ID will be shown. Furthermore, from now on the pdf id will be used
instead of the uuid.

fixes #47