fix: handle problems with viewing updated pdfs
When editing pdfs in the viewer and saving them, sometimes the
old version is still shown in the viewer even though the backend
has the correct version. This is probably caused by the browser's
caching. This commit fixes the problem by adding a revision to the
serve views so that the browser is forced to refresh the pdf.

closes #76